add pynput to requirements.txt, rename .py files for consistency, apply reformat code and optimize imports to all .py files, minor changes to indentation in scripts
