+++ README.md
@@ -22,5 +22,5 @@ pip install -r requirements.txt
 ### 4. Execute tests
 for example, for the `client2.py` test, execute:
 ```shell script
-python3 client2.py
+python3 client-2.py
 ```
