+++ requirements.txt
@@ -1,2 +1,3 @@
 grpcio-tools==1.33.2
 grpcio==1.33.2
+pynput==1.7.1
